feat(javascript): load translations from new space:opened event
