+++ .changeset/sharp-pears-teach.md
@@ -1,0 +1,5 @@
+---
+'@flatfile/javascript': minor
+---
+
+Add internationalization support with detected browser language or given override language.
