+++ packages/javascript/package.json
@@ -34,7 +34,7 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-json": "^6.0.1",
+    "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-sucrase": "^5.0.2",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-url": "^8.0.2",
@@ -54,6 +54,8 @@
   "dependencies": {
     "@flatfile/embedded-utils": "^1.3.0",
     "@flatfile/listener": "^1.0.4",
-    "@flatfile/plugin-record-hook": "^1.5.2"
+    "@flatfile/plugin-record-hook": "^1.5.2",
+    "i18next": "^23.11.5",
+    "i18next-browser-languagedetector": "^8.0.0"
   }
 }
